fixed setup
