+++ app/models/amenity_old.py
@@ -1,39 +1,39 @@
-import uuid
-from datetime import datetime
+# import uuid
+# from datetime import datetime
 
-class Amenity:
-    # id = None
-    # name = ""
-    # created_at = None
-    # updated_at = None
+# class Amenity:
+#     # id = None
+#     # name = ""
+#     # created_at = None
+#     # updated_at = None
 
-    def __init__(self, name):
-        if name is None:
-            raise ValueError("Required attributes not specified!")
+#     def __init__(self, name):
+#         if name is None:
+#             raise ValueError("Required attributes not specified!")
 
-        self.id = str(uuid.uuid4())
-        self.created_at = datetime.now()
-        self.updated_at = datetime.now()
-        self.name = name
+#         self.id = str(uuid.uuid4())
+#         self.created_at = datetime.now()
+#         self.updated_at = datetime.now()
+#         self.name = name
 
-    # --- Getters and Setters ---
-    @property
-    def name(self):
-        """ Returns value of property name """
-        return self._name
+#     # --- Getters and Setters ---
+#     @property
+#     def name(self):
+#         """ Returns value of property name """
+#         return self._name
 
-    @name.setter
-    def name(self, value):
-        """Setter for prop name"""
-        # ensure that the value is up to 50 characters after removing excess white-space
-        is_valid_name = 0 < len(value.strip()) <= 50
-        if is_valid_name:
-            self._name = value.strip()
-        else:
-            raise ValueError("Invalid name length!")
+#     @name.setter
+#     def name(self, value):
+#         """Setter for prop name"""
+#         # ensure that the value is up to 50 characters after removing excess white-space
+#         is_valid_name = 0 < len(value.strip()) <= 50
+#         if is_valid_name:
+#             self._name = value.strip()
+#         else:
+#             raise ValueError("Invalid name length!")
 
 
-    # --- Methods ---
-    def save(self):
-        """Update the updated_at timestamp whenever the object is modified"""
-        self.updated_at = datetime.now()
+#     # --- Methods ---
+#     def save(self):
+#         """Update the updated_at timestamp whenever the object is modified"""
+#         self.updated_at = datetime.now()
