--- app/models/baseclass.py
@@ -1,10 +1,0 @@
-from app import db
-import uuid
-from datetime import datetime
-
-class BaseModel(db.Model):
-    __abstract__ = True  # This ensures SQLAlchemy does not create a table for BaseModel
-
-    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
-    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
