+++ app/persistence/repository.py
@@ -1,4 +1,6 @@
 from abc import ABC, abstractmethod
+from app.persistence import db_session
+from app.models.user import User
 
 class Repository(ABC):
     @abstractmethod
@@ -56,3 +58,37 @@ class InMemoryRepository(Repository):
 
     def get_by_attribute(self, attr_name, attr_value):
         return next((obj for obj in self._storage.values() if getattr(obj, attr_name) == attr_value), None)
+
+class SQLAlchemyRepository(Repository):
+    def __init__(self, model):
+        self.model = model
+
+    def add(self, obj):
+        db_session.add(obj)
+        db_session.commit()
+
+    def get(self, obj_id):
+        return db_session.query(self.model).get(obj_id)
+
+    def get_all(self):
+        return db_session.query(self.model).all()
+
+    # added new method for getting email
+    # def get_user_by_email(self, email):
+    #     return super().get_by_attribute("_email", email)
+
+    def update(self, obj_id, data):
+        obj = self.get(obj_id)
+        if obj:
+            for key, value in data.items():
+                setattr(obj, key, value)
+            db_session.commit()
+
+    def delete(self, obj_id):
+        obj = self.get(obj_id)
+        if obj:
+            db_session.delete(obj)
+            db_session.commit()
+
+    def get_by_attribute(self, attr_name, attr_value):
+        return db_session.query(self.model).where(getattr(self.model, attr_name) == attr_value).first()
